backend/shared/requirements.txt
# Shared dependencies for all CloudCare API servers
fastapi==0.109.0
uvicorn[standard]==0.27.0
prisma==0.11.0
pydantic>=2.5.0,<3.0.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Utilities
python-dateutil==2.8.2
pytz==2023.3
